+++ mltu/__init__.py
@@ -1,3 +1,3 @@
-__version__ = "1.0.3"
+__version__ = "1.0.4"
 
 from .annotations.image import Image